Changed ids to Long type

+++ src/main/scala/dto/entities/PostDto.scala
@@ -3,4 +3,4 @@ package dto.entities
 import java.sql.Timestamp
 
 
-case class PostDto(content: String, secretKey: String, postTimestamp: Timestamp, userId: Int, topicId: Int, postId: Option[Int] = None)
+case class PostDto(content: String, secretKey: String, postTimestamp: Timestamp, userId: Long, topicId: Long, postId: Option[Long] = None)
