+++ src/main/scala/dto/entities/UserDto.scala
@@ -1,3 +1,3 @@
 package dto.entities
 
-case class UserDto(nickname: String, email: String, userId: Option[Int] = None)
+case class UserDto(nickname: String, email: String, userId: Option[Long] = None)
